Refactored resolvers to use an interface
* Spark\Resolver\ResolverInterface is now used for typehinting resolvers
* Spark\Resolver is now Spark\Resolver\AurynResolver
