+++ src/Handler/ActionHandler.php
@@ -10,18 +10,18 @@ use Psr\Http\Message\ResponseInterface;
 use Spark\Adr\DomainInterface;
 use Spark\Adr\InputInterface;
 use Spark\Adr\ResponderInterface;
-use Spark\Resolver;
+use Spark\Resolver\ResolverInterface;
 
 class ActionHandler extends Arbiter
 {
     /**
-     * @var Resolver
+     * @var Spark\Resolver\ResolverInterface
      */
     protected $resolver;
 
     protected $actionAttribute = 'spark/adr:action';
 
-    public function __construct(Resolver $resolver)
+    public function __construct(ResolverInterface $resolver)
     {
         $this->resolver = $resolver;
     }
