+++ src/Resolver/ResolverInterface.php
@@ -1,0 +1,11 @@
+<?php
+namespace Spark\Resolver;
+
+interface ResolverInterface
+{
+    /**
+     * @param mixed $spec
+     * @return callable
+     */
+    public function __invoke($spec);
+}
